+++ ui/routenodelabel.h
@@ -9,7 +9,7 @@ class RouteNodeLabel : public QGraphicsSimpleTextItem, public ViewCustomizable
 public:
     RouteNodeLabel(QString text);
 
-    virtual void setElementSize(qreal newSize) override;
+    virtual void setElementSize(int newSize) override;
     virtual void setColors(const QColor &color) override;
     virtual void setDefaultColors() override;
